+++ src/day_02.rs
@@ -1,20 +1,20 @@
 use std::cmp::{max, min};
 
-fn parse_input(input: String) -> Vec<Vec<u32>> {
+fn parse_input(input: String) -> Vec<Vec<u64>> {
     input
         .lines()
-        .map(|l| l.split(" ").map(|x| x.parse::<u32>().unwrap()).collect())
+        .map(|l| l.split(" ").map(|x| x.parse::<u64>().unwrap()).collect())
         .collect()
 }
 
-fn is_safe(report: &[u32]) -> bool {
+fn is_safe(report: &[u64]) -> bool {
     (report.iter().is_sorted_by(|a, b| a < b) || report.iter().is_sorted_by(|a, b| a > b))
         && report
             .windows(2)
             .all(|window| (max(window[0], window[1]) - min(window[0], window[1]) <= 3))
 }
 
-fn generate_subreports(report: &[u32]) -> Vec<Vec<u32>> {
+fn generate_subreports(report: &[u64]) -> Vec<Vec<u64>> {
     let mut subreports = Vec::new();
     subreports.push(report.to_owned());
     for ind in 0..report.len() {
@@ -25,12 +25,12 @@ fn generate_subreports(report: &[u32]) -> Vec<Vec<u32>> {
     subreports
 }
 
-pub fn part_one(input: String) -> u32 {
+pub fn part_one(input: String) -> u64 {
     let table = parse_input(input);
-    table.iter().filter(|report| is_safe(report)).count() as u32
+    table.iter().filter(|report| is_safe(report)).count() as u64
 }
 
-pub fn part_two(input: String) -> u32 {
+pub fn part_two(input: String) -> u64 {
     let table = parse_input(input);
     table
         .iter()
@@ -38,7 +38,7 @@ pub fn part_two(input: String) -> u32 {
             let subreports = generate_subreports(report);
             subreports.iter().any(|x| is_safe(x))
         })
-        .count() as u32
+        .count() as u64
 }
 
 #[cfg(test)]
